Fixed reload when prompt is not recognized
Since the program exitted the iframe it wasn't able to locate the refresh button on the captcha window and looped indefinitely on the same prompt.
Refreshing before switching to default content allows to locate the div and to click on it. Now prompts like "cat with a pirate hat" that aren't in the db are skipped and the bot converges more often towards a resolution.